0.6.7 Website - Refactor auth and unittest
