+++ .gitignore
@@ -1,3 +1,5 @@
 includes/
 flask_session/
 backupapp.py
+Website/databaseconn.py
+Website/photos
